Updated names of the converters
CollectionEmptiness -> EmptyCollection
StringEmptiness -> EmptyString
