+++ Microsoft.Toolkit.Uwp.UI/Converters/EmptyCollectionToObjectConverter.cs
@@ -1,7 +1,17 @@
-﻿using System;
+﻿// ******************************************************************
+// Copyright (c) Microsoft. All rights reserved.
+// This code is licensed under the MIT License (MIT).
+// THE CODE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+// INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+// IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+// DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
+// TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH
+// THE CODE OR THE USE OR OTHER DEALINGS IN THE CODE.
+// ******************************************************************
+using System;
 using System.Collections;
 using Windows.UI.Xaml.Data;
-using Windows.UI.Xaml.Markup;
 
 namespace Microsoft.Toolkit.Uwp.UI.Converters
 {
@@ -9,7 +19,7 @@ namespace Microsoft.Toolkit.Uwp.UI.Converters
     /// This class converts a collection size into an other object.
     /// Can be used to convert to bind a visibility, a color or an image to the size of the collection.
     /// </summary>
-    public class CollectionEmptinessToObjectConverter : IValueConverter
+    public class EmptyCollectionToObjectConverter : IValueConverter
     {
         /// <summary>
         /// Gets or sets the value to be returned when the collection is neither null nor empty
