+++ Microsoft.Toolkit.Uwp.UI/Converters/EmptyStringToObjectConverter.cs
@@ -1,6 +1,16 @@
-﻿using System;
+﻿// ******************************************************************
+// Copyright (c) Microsoft. All rights reserved.
+// This code is licensed under the MIT License (MIT).
+// THE CODE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+// INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+// IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+// DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
+// TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH
+// THE CODE OR THE USE OR OTHER DEALINGS IN THE CODE.
+// ******************************************************************
+using System;
 using Windows.UI.Xaml.Data;
-using Windows.UI.Xaml.Markup;
 
 namespace Microsoft.Toolkit.Uwp.UI.Converters
 {
@@ -8,7 +18,7 @@ namespace Microsoft.Toolkit.Uwp.UI.Converters
     /// This class converts a string value into a an object (if the value is null or empty returns the false value).
     /// Can be used to bind a visibility, a color or an image to the value of a string.
     /// </summary>
-    public class StringEmptinessToObjectConverter : IValueConverter
+    public class EmptyStringToObjectConverter : IValueConverter
     {
         /// <summary>
         /// Gets or sets the value to be returned when the string is neither null nor empty
