+++ Microsoft.Toolkit.Uwp.UI/Microsoft.Toolkit.Uwp.UI.csproj
@@ -51,12 +51,12 @@
   <ItemGroup>
     <Compile Include="Converters\BoolToObjectConverter.cs" />
     <Compile Include="Converters\BoolToVisibilityConverter.cs" />
-    <Compile Include="Converters\CollectionEmptinessToObjectConverter.cs" />
     <Compile Include="Converters\CollectionVisibilityConverter.cs" />
     <Compile Include="Converters\ConverterTools.cs" />
-    <Compile Include="Converters\StringEmptinessToObjectConverter.cs" />
-    <Compile Include="Converters\StringFormatConverter.cs" />
+    <Compile Include="Converters\EmptyCollectionToObjectConverter.cs" />
+    <Compile Include="Converters\EmptyStringToObjectConverter.cs" />
     <Compile Include="Converters\StringVisibilityConverter.cs" />
+    <Compile Include="Converters\StringFormatConverter.cs" />
     <Compile Include="ImageCache\ImageCache.cs" />
     <Compile Include="Extensions\VisualTreeExtensions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
