+++ UnitTests/UnitTests.csproj
@@ -114,8 +114,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Converters\Test_BoolToObjectConverter.cs" />
-    <Compile Include="Converters\Test_StringEmptinessToObjectConverter.cs" />
-    <Compile Include="Converters\Test_CollectionEmptinessToObjectConverter.cs" />
+    <Compile Include="Converters\Test_EmptyCollectionToObjectConverter.cs" />
+    <Compile Include="Converters\Test_EmptyStringToObjectConverter.cs" />
     <Compile Include="Helpers\Test_ColorHelper.cs" />
     <Compile Include="Helpers\Test_ConnectionHelper.cs" />
     <Compile Include="Helpers\Test_StorageFileHelper.cs" />
